Refactor (MCP): Refactor the packet structure of MCP server related parameter classes
-Move the McpError, McpErroCode, McpException, McpRequest, McpResponse, and ServerStatus classes to the param sub package
-Update the relevant import paths to reflect the new package structure
-In AbstractMcpConnection、McpClient、McpConnection、McpHttpConnection、McpServerConnection Add missing McpResponse import to McpSseConnection - Add Apache license header to IteratorAction class
