+++ src/main/java/org/salt/jlangchain/rag/tools/mcp/server/McpHttpConnection.java
@@ -17,6 +17,9 @@ package org.salt.jlangchain.rag.tools.mcp.server;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.*;
 import org.salt.jlangchain.rag.tools.mcp.server.config.ServerConfig;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpException;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpRequest;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpResponse;
 
 import java.io.BufferedReader;
 import java.io.IOException;
