+++ src/main/java/org/salt/jlangchain/rag/tools/mcp/server/McpServerConnection.java
@@ -2,6 +2,9 @@ package org.salt.jlangchain.rag.tools.mcp.server;
 
 import lombok.extern.slf4j.Slf4j;
 import org.salt.jlangchain.rag.tools.mcp.server.config.ServerConfig;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpException;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpRequest;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpResponse;
 
 import java.io.*;
 import java.util.ArrayList;
