+++ src/main/java/org/salt/jlangchain/rag/tools/mcp/server/McpSseConnection.java
@@ -21,6 +21,9 @@ import okhttp3.sse.EventSourceListener;
 import okhttp3.sse.EventSources;
 import org.jetbrains.annotations.NotNull;
 import org.salt.jlangchain.rag.tools.mcp.server.config.ServerConfig;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpException;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpRequest;
+import org.salt.jlangchain.rag.tools.mcp.server.param.McpResponse;
 
 import java.io.IOException;
 import java.util.HashMap;
